+++ th04/mem.h
@@ -2,7 +2,7 @@
 #define MEM_ASSIGN_PARAS_OP    (336000 >> 4)
 #define MEM_ASSIGN_PARAS_MAIN  (320000 >> 4)
 #define MEM_ASSIGN_PARAS_MAINE (336000 >> 4)
-extern unsigned int mem_assign_paras;
+extern size_t mem_assign_paras;
 
 #define EMSSIZE (180000) /* ZUN symbol [MAGNet2010] */
